refactor: inspector sidebar to a independent folder
