+++ packages/react-router-v7-example/app/components/inspector-sidebar/inspectPanel.tsx
@@ -1,0 +1,3 @@
+export function InspectPanel() {
+  return <div>Inspect</div>;
+}
