+++ packages/react-router-v7-example/app/components/inspector-sidebar/stylesPanel.tsx
@@ -1,0 +1,3 @@
+export function StylesPanel() {
+  return <div>Styles</div>;
+}
